+++ modules/aws-http4s/src/smithy4s/aws/internals/AwsRestXmlCodecs.scala
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2021-2025 Disney Streaming
+ *  Copyright 2021-2026 Disney Streaming
  *
  *  Licensed under the Tomorrow Open Source Technology License, Version 1.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -45,6 +45,7 @@ private[aws] object AwsRestXmlCodecs {
       .withRawStringsAndBlobsPayloads
       .withSmithyPathEncoding(true)
       .withRequestMediaType("application/xml")
+      .withAcceptMediaType("application/xml")
   }
 
 }