+++ modules/bootstrapped/test/src/smithy4s/schematests/SchemaPartitionSpec.scala
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2021-2025 Disney Streaming
+ *  Copyright 2021-2026 Disney Streaming
  *
  *  Licensed under the Tomorrow Open Source Technology License, Version 1.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -15,11 +15,13 @@
  */
 
 package smithy4s
-package schema
+package schematests
 
 import cats.syntax.all._
 import Schema._
 import munit._
+import smithy4s.schema.SchemaPartition
+import smithy4s.schema.Field
 
 final class SchemaPartitionSpec extends FunSuite {