+++ modules/protocol/src/smithy4s/meta/GenerateServiceProductTrait.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2021-2025 Disney Streaming
+ *  Copyright 2021-2026 Disney Streaming
  *
  *  Licensed under the Tomorrow Open Source Technology License, Version 1.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -25,7 +25,7 @@ import software.amazon.smithy.model.traits.AnnotationTrait;
 import software.amazon.smithy.model.traits.AbstractTrait;
 
 public class GenerateServiceProductTrait extends AnnotationTrait {
-  public static ShapeId ID = ShapeId.from("smithy4s.meta#generateServiceProduct");
+	public static ShapeId ID = ShapeId.from("smithy4s.meta#generateServiceProduct");
 
 	public GenerateServiceProductTrait(ObjectNode node) {
 		super(ID, node);